smoothen mouse position by default
