+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export * from "./use-orakul";
-export * from "./use-mouse-position";
+export * from "./use-smooth-mouse-position";
 export * from "./use-rpc-status";
 export * from "./use-notifications";
